workerの数を1に固定 e2eテスト実行前後のスクリプトでエラーが出ないように修正 chromium以外のブラウザを追加

+++ e2e/after-e2e-test.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
 pid=$(ps aux | grep 'dotnet run --project ../src/AspNetCoreSample.Mvc' | grep -v grep | awk '{print $2}')
-kill $pid
+if [ -n "$pid" ]; then
+    kill $pid
+fi
